OpenGL ES 3 renderer backend: Visual Studio project: Removed usage of "OPENGLES3RENDERER_NO_STATE_CLEANUP" since sadly the OpenGL ES 3 API is pretty bind oriented making it really difficult to keep state changes over control

+++ Renderer/RendererBackend/OpenGLES3Renderer/src/Shader/ProgramGlsl.cpp
@@ -122,7 +122,7 @@ namespace OpenGLES3Renderer
 										// Backup the currently used OpenGL ES 3 program
 										GLint openGLES3ProgramBackup = 0;
 										glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-										if (openGLES3ProgramBackup == mOpenGLES3Program)
+										if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 										{
 											// Set uniform, please note that for this our program must be the currently used one
 											glUniform1i(uniformLocation, static_cast<GLint>(descriptorRange->baseShaderRegister));
@@ -134,7 +134,7 @@ namespace OpenGLES3Renderer
 											glUniform1i(uniformLocation, static_cast<GLint>(descriptorRange->baseShaderRegister));
 
 											// Be polite and restore the previous used OpenGL ES 3 program
-											glUseProgram(openGLES3ProgramBackup);
+											glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 										}
 									#else
 										// Set uniform, please note that for this our program must be the currently used one
@@ -201,7 +201,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniform1i(static_cast<GLint>(uniformHandle), value);
@@ -213,7 +213,7 @@ namespace OpenGLES3Renderer
 				glUniform1i(static_cast<GLint>(uniformHandle), value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -228,7 +228,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniform1f(static_cast<GLint>(uniformHandle), value);
@@ -240,7 +240,7 @@ namespace OpenGLES3Renderer
 				glUniform1f(static_cast<GLint>(uniformHandle), value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -255,7 +255,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniform2fv(static_cast<GLint>(uniformHandle), 1, value);
@@ -267,7 +267,7 @@ namespace OpenGLES3Renderer
 				glUniform2fv(static_cast<GLint>(uniformHandle), 1, value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -282,7 +282,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniform3fv(static_cast<GLint>(uniformHandle), 1, value);
@@ -294,7 +294,7 @@ namespace OpenGLES3Renderer
 				glUniform3fv(static_cast<GLint>(uniformHandle), 1, value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -309,7 +309,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniform4fv(static_cast<GLint>(uniformHandle), 1, value);
@@ -321,7 +321,7 @@ namespace OpenGLES3Renderer
 				glUniform4fv(static_cast<GLint>(uniformHandle), 1, value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -336,7 +336,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniformMatrix3fv(static_cast<GLint>(uniformHandle), 1, GL_FALSE, value);
@@ -348,7 +348,7 @@ namespace OpenGLES3Renderer
 				glUniformMatrix3fv(static_cast<GLint>(uniformHandle), 1, GL_FALSE, value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
@@ -363,7 +363,7 @@ namespace OpenGLES3Renderer
 			// Backup the currently used OpenGL ES 3 program
 			GLint openGLES3ProgramBackup = 0;
 			glGetIntegerv(GL_CURRENT_PROGRAM, &openGLES3ProgramBackup);
-			if (openGLES3ProgramBackup == mOpenGLES3Program)
+			if (openGLES3ProgramBackup == static_cast<GLint>(mOpenGLES3Program))
 			{
 				// Set uniform, please note that for this our program must be the currently used one
 				glUniformMatrix4fv(static_cast<GLint>(uniformHandle), 1, GL_FALSE, value);
@@ -375,7 +375,7 @@ namespace OpenGLES3Renderer
 				glUniformMatrix4fv(static_cast<GLint>(uniformHandle), 1, GL_FALSE, value);
 
 				// Be polite and restore the previous used OpenGL ES 3 program
-				glUseProgram(openGLES3ProgramBackup);
+				glUseProgram(static_cast<GLuint>(openGLES3ProgramBackup));
 			}
 		#else
 			// Set uniform, please note that for this our program must be the currently used one
